0.2.0

+++ openhrm/routes/position.route.js
@@ -142,15 +142,27 @@ route.delete('/:id', (req, res) => {
 
     let positionId = req.params.id;
 
-    db.remove(Position, { _id: positionId })
-        .then(() => {
+  Position.count({},( err, count) =>{
+    if(count > 1){
+        db.remove(Position, {_id: positionId})
+          .then(() => {
             res.statusCode = statusCode.deleted;
             res.send();
-        })
-        .catch(() => {
+          })
+          .catch(() => {
             res.statusCode = statusCode.notfound;
             res.send();
+          });
+    }else{
+        res.statusCode = statusCode.bad;
+        res.json({
+          error: "You have to have at least one default position to make all the things up and runing!"
         });
+    }
+  })
+
+
+
 });
 
 module.exports = route;
